ggml-cuda: fixed assertion in ggml_cuda_cpy (#18140)

+++ ggml/src/ggml-cuda/cpy.cu
@@ -356,8 +356,8 @@ void ggml_cuda_cpy(ggml_backend_cuda_context & ctx, const ggml_tensor * src0, gg
     const int64_t ne = ggml_nelements(src0);
     GGML_ASSERT(ne == ggml_nelements(src1));
 
-    GGML_ASSERT(ggml_nbytes(src0) <= INT_MAX);
-    GGML_ASSERT(ggml_nbytes(src1) <= INT_MAX);
+    GGML_ASSERT(ggml_nbytes(src0) <= SIZE_MAX);
+    GGML_ASSERT(ggml_nbytes(src1) <= SIZE_MAX);
 
     const int64_t ne00 = src0->ne[0];
     const int64_t ne01 = src0->ne[1];
